emg-functions-ar-billing-dev/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
numpy
pandas
matplotlib
fpdf2
SQLAlchemy
python-dotenv
pyodbc
